requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pbr==5.4.2
PyJWT==1.7.1
PySocks==1.7.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
stevedore==1.30.1
twilio==6.29.3
urllib3==1.25.3
virtualenv==16.7.2
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
Werkzeug==0.15.5
